# backend/api-spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PodAsk API",
        "description": "Transform scientific papers into interactive podcasts",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/podcasts/generate": {
            "post": {
                "tags": [
                    "podcasts"
                ],
                "summary": "Generate Podcast",
                "description": "Start podcast generation from ingested paper IDs. Papers must be ingested first via POST /papers/ingest. Returns immediately, generation happens in background. Use GET /{podcast_id}/status to poll for completion.",
                "operationId": "generate_podcast_api_v1_podcasts_generate_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PodcastGenerateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PodcastResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/podcasts": {
            "get": {
                "tags": [
                    "podcasts"
                ],
                "summary": "List Podcasts",
                "description": "List all podcasts for the current user.",
                "operationId": "list_podcasts_api_v1_podcasts_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PodcastListResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/podcasts/{podcast_id}": {
            "get": {
                "tags": [
                    "podcasts"
                ],
                "summary": "Get Podcast",
                "description": "Get podcast with all segments.",
                "operationId": "get_podcast_api_v1_podcasts__podcast_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "podcast_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Podcast Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PodcastResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "podcasts"
                ],
                "summary": "Delete Podcast",
                "description": "Delete a podcast and its segments.",
                "operationId": "delete_podcast_api_v1_podcasts__podcast_id__delete",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "podcast_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Podcast Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/podcasts/{podcast_id}/status": {
            "get": {
                "tags": [
                    "podcasts"
                ],
                "summary": "Get Podcast Status",
                "description": "Get podcast generation status (for polling).",
                "operationId": "get_podcast_status_api_v1_podcasts__podcast_id__status_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "podcast_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Podcast Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PodcastStatusResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/podcasts/{podcast_id}/audio/{segment_sequence}": {
            "get": {
                "tags": [
                    "podcasts"
                ],
                "summary": "Get Segment Audio",
                "description": "Stream audio for a specific segment.",
                "operationId": "get_segment_audio_api_v1_podcasts__podcast_id__audio__segment_sequence__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "podcast_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Podcast Id"
                        }
                    },
                    {
                        "name": "segment_sequence",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Segment Sequence"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DialogueLine": {
                "properties": {
                    "speaker": {
                        "type": "string",
                        "title": "Speaker"
                    },
                    "text": {
                        "type": "string",
                        "title": "Text"
                    },
                    "audio_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Audio Url"
                    }
                },
                "type": "object",
                "required": [
                    "speaker",
                    "text"
                ],
                "title": "DialogueLine"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PodcastGenerateRequest": {
                "properties": {
                    "paper_ids": {
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "type": "array",
                        "title": "Paper Ids",
                        "description": "List of ingested paper UUIDs to synthesize into a podcast. Papers must be ingested first via POST /papers/ingest.",
                        "minItems": 1,
                        "maxItems": 10
                    },
                    "topic": {
                        "type": "string",
                        "title": "Topic",
                        "description": "Topic/context for the podcast. This will guide the narrative focus."
                    },
                    "difficulty_level": {
                        "type": "string",
                        "title": "Difficulty Level",
                        "description": "Target audience difficulty level: beginner, intermediate, or advanced.",
                        "default": "intermediate",
                        "enum": ["beginner", "intermediate", "advanced"]
                    }
                },
                "type": "object",
                "required": [
                    "paper_ids",
                    "topic"
                ],
                "title": "PodcastGenerateRequest"
            },
            "PodcastListResponse": {
                "properties": {
                    "podcasts": {
                        "items": {
                            "$ref": "#/components/schemas/PodcastResponse"
                        },
                        "type": "array",
                        "title": "Podcasts"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    }
                },
                "type": "object",
                "required": [
                    "podcasts",
                    "total"
                ],
                "title": "PodcastListResponse"
            },
            "PodcastResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id",
                        "description": "Unique podcast identifier"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title",
                        "description": "Generated podcast title"
                    },
                    "summary": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Summary",
                        "description": "Generated podcast summary"
                    },
                    "topic": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Topic",
                        "description": "Original topic provided for generation"
                    },
                    "paper_ids": {
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "type": "array",
                        "title": "Paper Ids",
                        "description": "List of source paper UUIDs used to generate this podcast"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "total_duration_seconds": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Total Duration Seconds"
                    },
                    "error_message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Message"
                    },
                    "created_at": {
                        "type": "string",
                        "title": "Created At"
                    },
                    "segments": {
                        "items": {
                            "$ref": "#/components/schemas/SegmentResponse"
                        },
                        "type": "array",
                        "title": "Segments",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "paper_ids",
                    "status",
                    "created_at"
                ],
                "title": "PodcastResponse"
            },
            "PodcastStatusResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "error_message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Message"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "status"
                ],
                "title": "PodcastStatusResponse"
            },
...
            "SegmentResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "sequence": {
                        "type": "integer",
                        "title": "Sequence"
                    },
                    "topic_label": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Topic Label"
                    },
                    "dialogue": {
                        "items": {
                            "$ref": "#/components/schemas/DialogueLine"
                        },
                        "type": "array",
                        "title": "Dialogue"
                    },
                    "key_terms": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Key Terms",
                        "default": []
                    },
                    "difficulty_level": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Difficulty Level"
                    },
                    "audio_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Audio Url"
                    },
                    "duration_seconds": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Duration Seconds"
                    },
                    "transition_to_question": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Transition To Question"
                    },
                    "resume_phrase": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Resume Phrase"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "sequence",
                    "dialogue"
                ],
                "title": "SegmentResponse"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}